+++ main/src/mapgen/mapgen_internal.h
@@ -27,6 +27,9 @@ void place_secret_treasures(unsigned char treasure_count);
 // False corridor management
 void place_false_corridors(unsigned char corridor_count);
 
+// Hidden corridor management
+void place_hidden_corridors(unsigned char corridor_count);
+
 // Corridor breakpoint management
 void calculate_and_store_breakpoints(unsigned char room_idx, unsigned char connection_idx);